+++ src/ReactPixiFiber.js
@@ -120,7 +120,7 @@ function render(pixiElement, stage, callback) {
   ReactPixiFiber.injectIntoDevTools({
     findFiberByHostInstance: ReactPixiFiber.findFiberByHostInstance,
     bundleType: 1,
-    version: "0.3.1",
+    version: "0.4.0",
     rendererPackageName: "react-pixi-fiber",
   });
 }